Phase 4: Kitty

+++ tabs/applications-setup/kitty-setup.sh
@@ -3,25 +3,38 @@
 . ../common-script.sh
 
 installKitty() {
-    printf "%b\n" "${YELLOW}Install Kitty if not already installed...${RC}"
-    if ! command_exists kitty; then
-        case "$PACKAGER" in
-            pacman)
-                "$ESCALATION_TOOL" "$PACKAGER" -S --needed --noconfirm kitty
-                ;;
-            *)
-                "$ESCALATION_TOOL" "$PACKAGER" install -y kitty
-                ;;
-        esac
+    printf "%b" "${YELLOW}Do you want to install Kitty? (Y/n): ${RC}"
+    read -r install_choice
+    if [ "$install_choice" != "n" ] && [ "$install_choice" != "N" ]; then
+        printf "%b\n" "${YELLOW}Installing Kitty...${RC}"
+        if ! command_exists kitty; then
+            case "$PACKAGER" in
+                pacman)
+                    "$ESCALATION_TOOL" "$PACKAGER" -S --needed --noconfirm kitty
+                    ;;
+                *)
+                    "$ESCALATION_TOOL" "$PACKAGER" install -y kitty
+                    ;;
+            esac
+        else
+            printf "%b\n" "${GREEN}Kitty is already installed.${RC}"
+        fi
     else
-        printf "%b\n" "${GREEN}Kitty is already installed.${RC}"
+        printf "%b\n" "${GREEN}Skipping Kitty installation.${RC}"
     fi
 }
 
 setupKittyConfig() {
-    printf "%b\n" "${YELLOW}Copying Kitty configuration files...${RC}"
-    if [ -d "${HOME}/.config/kitty" ] && [ ! -d "${HOME}/.config/kitty-bak" ]; then
-        cp -r "${HOME}/.config/kitty" "${HOME}/.config/kitty-bak"
+    printf "%b" "${YELLOW}Do you want to backup existing Kitty configuration files? (Y/n): ${RC}"
+    read -r backup_choice
+    if [ "$backup_choice" != "n" ] && [ "$backup_choice" != "N" ]; then
+        printf "%b\n" "${YELLOW}Backing up existing Kitty configuration files...${RC}"
+        if [ -d "${HOME}/.config/kitty" ] && [ ! -d "${HOME}/.config/kitty-bak" ]; then
+            cp -r "${HOME}/.config/kitty" "${HOME}/.config/kitty-bak"
+        fi
+        printf "%b\n" "${GREEN}Kitty configuration files backed up.${RC}"
+    else
+        printf "%b\n" "${GREEN}Skipping backup of Kitty configuration files.${RC}"
     fi
     mkdir -p "${HOME}/.config/kitty/"
     curl -sSLo "${HOME}/.config/kitty/kitty.conf" https://github.com/ChrisTitusTech/dwm-titus/raw/main/config/kitty/kitty.conf
